requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==23.1.0
base58==2.1.1
behave==1.2.6
bitarray==2.7.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coincurve==18.0.0
cytoolz==0.12.1
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
exceptiongroup==1.1.1
frozenlist==1.3.3
greenlet==2.0.2
hexbytes==0.3.0
idna==3.4
iniconfig==2.0.0
ipfshttpclient==0.8.0a2
jsonschema==4.17.3
lru-dict==1.1.8
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
packaging==23.1
parse==1.19.0
parse-type==0.6.0
parsimonious==0.8.1
pluggy==1.0.0
protobuf==3.20.3
py-solc-x==1.1.1
pycparser==2.21
pycryptodome==3.17
pyrsistent==0.19.3
pysha3==1.0.2
pytest==7.3.1
regex==2023.3.23
requests==2.28.2
rlp==2.0.1
semantic-version==2.10.0
six==1.16.0
solc==0.0.0a0
SQLAlchemy==2.0.10
tomli==2.0.1
toolz==0.12.0
typing_extensions==4.5.0
urllib3==1.26.15
varint==1.0.2
web3==5.29.2
websockets==9.1
yarl==1.9.1
